# Improve performance of `add_row_number`