--- a/distribution/lib/Standard/Table/0.0.0-dev/src/Internal/Add_Row_Number.enso
+++ b/distribution/lib/Standard/Table/0.0.0-dev/src/Internal/Add_Row_Number.enso
@@ -14,6 +14,7 @@ from project.Internal.Java_Exports import make_long_builder
 polyglot java import org.enso.base.arrays.MutableLongArrayWrapper
 polyglot java import org.enso.table.data.column.storage.numeric.LongRangeStorage
 polyglot java import org.enso.table.data.column.storage.numeric.LongStorage
+polyglot java import org.enso.table.operations.AddRowNumber
 
 ## PRIVATE
 add_row_number : Table -> Text -> Integer -> Integer -> Text | Integer | Regex | Vector (Integer | Text | Regex) -> Vector (Text | Sort_Column) | Text | Sort_Column -> Problem_Behavior -> Table
@@ -59,46 +60,24 @@ make_range_column name start step length =
 
 ## PRIVATE
 make_grouped_enumeration name grouping_columns start step =
-    n = grouping_columns.at 0 . length
-    column_builder = make_long_builder n
-    0.up_to n . fold Map.empty grouping_counters-> ix->
-        key = Unordered_Multi_Value_Key.from_row grouping_columns ix
-        enum_index = grouping_counters.get key 0
-        column_builder.appendLong (nth_index start step enum_index)
-        new_counters = grouping_counters.insert key (enum_index + 1)
-        new_counters
-    storage = column_builder.seal
-    Column.from_storage name storage
+    grouping_java_columns = grouping_columns.map col->
+        col.java_column
+    new_storage = AddRowNumber.create_grouped_numbering start step grouping_java_columns
+    Column.from_storage name new_storage
 
 ## PRIVATE
    If the `grouping_columns` are empty, all rows are considered to be in the same group.
 make_grouped_ordered_enumeration name grouping_columns ordering from step =
-    ordering_columns = ordering.map .column
-    ordering_flip_directions = ordering.map descriptor-> case descriptor.associated_selector.direction of
-        Sort_Direction.Ascending -> False
-        Sort_Direction.Descending -> True
-    n = ordering_columns.at 0 . length
-    grouped_rows = (0.up_to n).fold Map.empty grouped_rows-> ix->
-        key = Unordered_Multi_Value_Key.from_row grouping_columns ix
-        new_grouped_rows = case grouped_rows.get key of
-            Nothing ->
-                builder = Vector.new_builder
-                builder.append ix
-                grouped_rows.insert key builder
-            existing_builder ->
-                existing_builder.append ix
-                grouped_rows
-        new_grouped_rows
+    ordering_columns = ordering.map c->
+        c.column.java_column
+    directions = ordering.map c->
+        c.associated_selector.direction.to_sign
 
-    long_array = MutableLongArrayWrapper.new n
+    grouping_java_columns = grouping_columns.map col-> col.java_column
 
-    grouped_rows.each row_group_builder->
-        row_group = row_group_builder.to_vector
-        sorted_group = row_group.sort on=ix->
-            Ordered_Multi_Value_Key.from_row ordering_columns ordering_flip_directions ix
-        sorted_group.each_with_index enum_ix-> row_ix->
-            enum_value = nth_index from step enum_ix
-            long_array.set row_ix enum_value
-
-    storage = LongStorage.fromArray long_array.getUnderlyingArray
-    Column.from_storage name storage
+    new_storage = case grouping_java_columns.is_empty of
+        True ->
+            AddRowNumber.create_ordered_numbering from step ordering_columns directions
+        False ->
+            AddRowNumber.create_grouped_ordered_numbering from step ordering_columns directions grouping_java_columns
+    Column.from_storage name new_storage