--- a/distribution/lib/Standard/Table/0.0.0-dev/src/Internal/Add_Row_Number.enso
+++ b/distribution/lib/Standard/Table/0.0.0-dev/src/Internal/Add_Row_Number.enso
@@ -1,4 +1,6 @@
 from Standard.Base import all
+import Standard.Base.Errors.Common.Unsupported_Argument_Types
+import Standard.Base.Errors.Illegal_Argument.Illegal_Argument
 
 import project.Data.Column.Column
 import project.Data.Set_Mode.Set_Mode
@@ -28,7 +30,7 @@ add_row_number table name from step group_by order_by on_problems =
     handle_arithmetic_exception _ =
         Error.throw (Illegal_Argument.Error "The row number has exceeded the 64-bit integer range. BigInteger numbering is currently not supported. Please use a smaller start/step.")
 
-    problem_builder.attach_problems_before on_problems <| Panic.catch ArithmeticException handle_arithmetic_exception <|
+    problem_builder.attach_problems_before on_problems <| Panic.catch ArithmeticException handler=handle_arithmetic_exception <| Panic.catch Unsupported_Argument_Types handler=handle_arithmetic_exception <|
         new_column = case ordering.is_empty of
             True ->
                 case grouping_columns.is_empty of