--- a/distribution/lib/Standard/Table/0.0.0-dev/src/Internal/Add_Row_Number.enso
+++ b/distribution/lib/Standard/Table/0.0.0-dev/src/Internal/Add_Row_Number.enso
@@ -67,8 +67,7 @@ make_range_column name start step length =
 
 ## PRIVATE
 make_grouped_enumeration name grouping_columns start step =
-    grouping_java_columns = grouping_columns.map col->
-        col.java_column
+    grouping_java_columns = grouping_columns.map .java_column
     new_storage = AddRowNumber.create_grouped_numbering start step grouping_java_columns
     Column.from_storage name new_storage
 
@@ -80,7 +79,7 @@ make_grouped_ordered_enumeration name grouping_columns ordering from step =
     directions = ordering.map c->
         c.associated_selector.direction.to_sign
 
-    grouping_java_columns = grouping_columns.map col-> col.java_column
+    grouping_java_columns = grouping_columns.map .java_column
 
     new_storage = case grouping_java_columns.is_empty of
         True ->