--- a/std-bits/table/src/main/java/org/enso/table/data/column/storage/numeric/LongRangeStorage.java
+++ b/std-bits/table/src/main/java/org/enso/table/data/column/storage/numeric/LongRangeStorage.java
@@ -12,6 +12,14 @@ public class LongRangeStorage extends ComputedLongStorage {
     super(size);
     this.start = start;
     this.step = step;
+    verifyBounds();
+  }
+
+  @SuppressWarnings("unused")
+  private void verifyBounds() throws ArithmeticException {
+    long lastIdx = size - 1;
+    // Computing this value will throw an exception if it overflows.
+    long lastValue = Math.addExact(start, Math.multiplyExact(step, lastIdx));
   }
 
   @Override