… test/Benchmarks/src/Main.enso
… test/Benchmarks/src/Table/Add_Row_Number.enso
… test/Table_Tests/src/Common_Table_Operations/Add_Row_Number_Spec.enso