@radeusgd 2023-10-16
Comparing the added benchmark before switching to the Java implementation:

| Benchmark | Old impl (Enso) | New impl (Java) |
| --- | ---: | ---: |
| Simple | 0.098 ms | 0.059 |
| Grouped | 3221.116 ms | 3.329 ms |
| Ordered | 2050.130 ms | 36.468 ms |
| Grouped+Ordered | 1672.881 ms | 19.377 ms |
| _Sort_Table_ | 8.658 ms | 4.722 ms |
| _Group and sort_ | 8.987 ms | 8.407 ms |

Explanation:
- `Simple` means an `add_row_number` without any grouping or ordering - here the implementation did not really change.
- `Sort_Table` is a baseline comparison of `table.order_by`
- `Group_And_Sort` is a baseline comparison of `table.aggregate `[Group_By "X", First "Y" order_by=["Y"]]` which causes a grouping and sorting similar to `add_row_number group_by="X" order_by="Y"`.

Summary:
We can see that the old implementation was magnitudes slower than the 'baseline' table sort or table aggregation. The new implementation is +- on par with the baseline.

<details>
<summary>Benchmark data</summary>

Before:
```
Label,Phase,"Invocations count","Average time (ms)"
"Add_Row_Number_50000.Simple","Warmup",8865,0.451
"Add_Row_Number_50000.Simple","Measurement",61183,0.098
"Add_Row_Number_50000.Grouped","Warmup",1,4738.558
"Add_Row_Number_50000.Grouped","Measurement",2,3221.116
"Add_Row_Number_50000.Ordered","Warmup",1,6120.888
"Add_Row_Number_50000.Ordered","Measurement",3,2050.13
"Add_Row_Number_50000.Grouped_And_Ordered","Warmup",3,1856.407
"Add_Row_Number_50000.Grouped_And_Ordered","Measurement",4,1672.881
"Add_Row_Number_50000.Sort_Table","Warmup",422,9.491
"Add_Row_Number_50000.Sort_Table","Measurement",694,8.658
"Add_Row_Number_50000.Group_And_Sort","Warmup",352,11.372
"Add_Row_Number_50000.Group_And_Sort","Measurement",668,8.987
```

After:
```
Label,Phase,"Invocations count","Average time (ms)"
"Add_Row_Number_50000.Simple","Warmup",26577,0.151
"Add_Row_Number_50000.Simple","Measurement",100861,0.059
"Add_Row_Number_50000.Grouped","Warmup",949,4.217
"Add_Row_Number_50000.Grouped","Measurement",1803,3.329
"Add_Row_Number_50000.Ordered","Warmup",99,40.556
"Add_Row_Number_50000.Ordered","Measurement",165,36.468
"Add_Row_Number_50000.Grouped_And_Ordered","Warmup",168,23.852
"Add_Row_Number_50000.Grouped_And_Ordered","Measurement",310,19.377
"Add_Row_Number_50000.Sort_Table","Warmup",712,5.628
"Add_Row_Number_50000.Sort_Table","Measurement",1272,4.722
"Add_Row_Number_50000.Group_And_Sort","Warmup",320,12.512
"Add_Row_Number_50000.Group_And_Sort","Measurement",714,8.407
```
</details>